Renamed the package and added setup file
