+++ .idea/runConfigurations/Server.xml
@@ -12,7 +12,7 @@
       <env name="PEPPER" value="hcQoBj+~Z#JxLd0K" />
     </envs>
     <option name="SDK_HOME" value="$USER_HOME$/.local/share/virtualenvs/backend-34xQ8ywB/bin/python" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/src" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/arduino_backend" />
     <option name="IS_MODULE_SDK" value="false" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
@@ -27,7 +27,7 @@
         <ENTRY IS_ENABLED="true" PARSER="runconfig" />
       </ENTRIES>
     </EXTENSION>
-    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/src/main.py" />
+    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/arduino_backend/__init__.py" />
     <option name="PARAMETERS" value="" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
