+++ arduino_backend/__main__.py
@@ -1,0 +1,4 @@
+import uvicorn
+from arduino_backend.main import app
+
+uvicorn.run(app, host="0.0.0.0", port=8000)
