+++ arduino_backend/main.py
@@ -1,11 +1,8 @@
 from fastapi import FastAPI
-from src.routers import user, auth
-import uvicorn
+
+from arduino_backend.routers import user, auth
 
 app = FastAPI()
 
 app.include_router(user.router, prefix="/user")
 app.include_router(auth.router, prefix="/auth")
-
-if __name__ == "__main__":
-    uvicorn.run(app, host="0.0.0.0", port=8000)
