+++ arduino_backend/routers/auth.py
@@ -4,16 +4,16 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 
-import src.models.user
-from src import database, settings
-from src.schemas import auth as authSchema
+import arduino_backend.models.user
+from arduino_backend import database, settings
+from arduino_backend.schemas import auth as authSchema
 
 router = APIRouter()
 
 
 @router.post("/", tags=["auth"], response_model=authSchema.Token)
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(database.get_db)):
-    user = src.models.user.User.authenticate_user(db=db, username=form_data.username, password=form_data.password)
+    user = arduino_backend.models.user.User.authenticate_user(db=db, username=form_data.username, password=form_data.password)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
@@ -21,7 +21,7 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
             headers={"WWW-Authenticate": "Bearer"},
         )
     access_token_expires = timedelta(minutes=settings.config.ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = src.models.user.User.create_access_token(
+    access_token = arduino_backend.models.user.User.create_access_token(
         data={"user": user.username}, expires_delta=access_token_expires
     )
     return {"access_token": access_token, "token_type": "bearer"}
