+++ arduino_backend/routers/user.py
@@ -1,17 +1,17 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
-import src.models.user
-from src import database
-from src.models import user as userModels
-from src.schemas import user as userSchema
+import arduino_backend.models.user
+from arduino_backend import database
+from arduino_backend.models import user as userModels
+from arduino_backend.schemas import user as userSchema
 
 router = APIRouter()
 
 
 @router.post("/", tags=["user"])
 def create_user(user: userSchema.User, db: Session = Depends(database.get_db)):
-    return src.models.user.User.create_user(db=db, user=user)
+    return arduino_backend.models.user.User.create_user(db=db, user=user)
 
 
 @router.get("/", tags=["user"])
@@ -21,7 +21,7 @@ def get_users(db: Session = Depends(database.get_db)):
 
 @router.get("/{username}", tags=["user"])
 def get_user_by_name(username: str, db: Session = Depends(database.get_db)):
-    user = src.models.user.User.get_user(db=db, username=username)
+    user = arduino_backend.models.user.User.get_user(db=db, username=username)
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
     else:
@@ -30,12 +30,12 @@ def get_user_by_name(username: str, db: Session = Depends(database.get_db)):
 
 @router.delete("/{username}", tags=["user"])
 def delete_user(username: str, db: Session = Depends(database.get_db)):
-    user = src.models.user.User.get_user(db=db, username=username)
+    user = arduino_backend.models.user.User.get_user(db=db, username=username)
     return user.delete(db=db)
 
 
 @router.put("/{username}", tags=["user"])
 def update_user(username: str, data: userSchema.User, db: Session = Depends(database.get_db)):
     partial_data = data.dict(exclude_unset=True)
-    user = src.models.user.User.get_user(db=db, username=username)
+    user = arduino_backend.models.user.User.get_user(db=db, username=username)
     return user.update_user(db=db, data=partial_data)
